requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.2
shap==0.44.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
openai==1.3.7
requests==2.31.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
jinja2==3.1.2
aiofiles==23.2.1
